#769 :recycle: move archcondition to separate file
